Move classmethod (from_* methods) and tests
